+++ backend/src/migrations.rs
@@ -266,15 +266,17 @@ pub enum SchemaVersion {
     CreatePltTokenAndEventTables,
     #[display("0040: Alter PLT events add event_timestamp and index")]
     AlterPltEventsAddEventTimestampAndIndex,
+    #[display("0041: m0041_fix_token_module_reject_reasons")]
+    FixTokenModuleRejectReasons,
 }
 impl SchemaVersion {
     /// The minimum supported database schema version for the API.
     /// Fails at startup if any breaking (destructive) database schema versions
     /// have been introduced since this version.
     pub const API_SUPPORTED_SCHEMA_VERSION: SchemaVersion =
-        SchemaVersion::CreatePltTokenAndEventTables;
+        SchemaVersion::AlterPltEventsAddEventTimestampAndIndex;
     /// The latest known version of the schema.
-    const LATEST: SchemaVersion = SchemaVersion::AlterPltEventsAddEventTimestampAndIndex;
+    const LATEST: SchemaVersion = SchemaVersion::FixTokenModuleRejectReasons;
 
     /// Parse version number into a database schema version.
     /// None if the version is unknown.
@@ -334,6 +336,7 @@ impl SchemaVersion {
             SchemaVersion::BakerApyQueryUpdateProtectAgainstOverflow => false,
             SchemaVersion::CreatePltTokenAndEventTables => false,
             SchemaVersion::AlterPltEventsAddEventTimestampAndIndex => false,
+            SchemaVersion::FixTokenModuleRejectReasons => false,
         }
     }
 
@@ -384,6 +387,7 @@ impl SchemaVersion {
             SchemaVersion::BakerApyQueryUpdateProtectAgainstOverflow => false,
             SchemaVersion::CreatePltTokenAndEventTables => false,
             SchemaVersion::AlterPltEventsAddEventTimestampAndIndex => false,
+            SchemaVersion::FixTokenModuleRejectReasons => false,
         }
     }
 
@@ -644,8 +648,16 @@ impl SchemaVersion {
                     .await?;
                 SchemaVersion::AlterPltEventsAddEventTimestampAndIndex
             }
+            SchemaVersion::AlterPltEventsAddEventTimestampAndIndex => {
+                tx.as_mut()
+                    .execute(sqlx::raw_sql(include_str!(
+                        "./migrations/m0041_fix_token_module_reject_reasons.sql"
+                    )))
+                    .await?;
+                SchemaVersion::FixTokenModuleRejectReasons
+            }
 
-            SchemaVersion::AlterPltEventsAddEventTimestampAndIndex => unimplemented!(
+            SchemaVersion::FixTokenModuleRejectReasons => unimplemented!(
                 "No migration implemented for database schema version {}",
                 self.as_i64()
             ),